requirements.txt
selenium==4.9.0
python-dotenv==0.21.0
